requirements.txt
asgiref==3.3.1
bcrypt==3.2.0
cffi==1.14.4
dj-database-url==0.5.0
Django==3.1.4
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
